Make edit functionality and refactor the code by puting the code in different folder
