+++ db.json
@@ -1,11 +1,12 @@
 {
   "todos": [
     {
-      "completed": true,
+      "completed": false,
       "todo": "Learn React",
       "id": "1"
     },
     {
+      "completed": false,
       "todo": "Learn Redux",
       "id": "2"
     },
@@ -13,6 +14,11 @@
       "completed": false,
       "todo": "Learn Redux-toolkit",
       "id": "1742995141095"
+    },
+    {
+      "completed": false,
+      "todo": "Learn Redux-toolkit with createAsyncThunk",
+      "id": "1742998656266"
     }
   ]
 }
