+++ src/features/todos/todoSlice.ts
@@ -1,23 +1,13 @@
-import { createSlice, createAsyncThunk } from '@reduxjs/toolkit';
-import { PayloadAction } from '@reduxjs/toolkit';
-import axios from '../../lib/axios';
-
-export interface Todo {
-  id: string;
-  text: string;
-  completed: boolean;
-}
-
-export interface TodoState {
-  todos: Todo[];
-  status: 'idle' | 'loading' | 'succeeded' | 'failed';
-  errors: {
-    [key: string]: string | null;
-  };
-  loadingStates: {
-    [key: string]: boolean;
-  };
-}
+import { createSlice, PayloadAction } from '@reduxjs/toolkit';
+import { Todo, TodoState } from './todoTypes';
+import {
+  fetchTodos,
+  addTodoAsync,
+  toggleTodoAsync,
+  deleteTodoAsync,
+  editTodoAsync,
+} from './todoThunks';
+import { handlePending, handleFulfilled, handleRejected } from './todoUtils';
 
 const initialState: TodoState = {
   todos: [],
@@ -25,97 +15,6 @@ const initialState: TodoState = {
   errors: {},
   loadingStates: {},
 };
-
-// Async thunks
-export const fetchTodos = createAsyncThunk('todos/fetchTodos', async (_, { rejectWithValue }) => {
-  try {
-    const response = await axios.get('/todos?limit=5');
-    return response.data.map((todo: any) => ({
-      id: todo.id.toString(),
-      text: todo.todo,
-      completed: todo.completed,
-    }));
-  } catch (error: any) {
-    return rejectWithValue(error.response?.data?.message || 'Failed to fetch todos');
-  }
-});
-
-export const addTodoAsync = createAsyncThunk(
-  'todos/addTodoAsync',
-  async (text: string, { rejectWithValue }) => {
-    try {
-      const response = await axios.post('/todos', {
-        todo: text,
-        completed: false,
-        id: Date.now().toString(),
-      });
-
-      return {
-        id: response.data.id.toString(),
-        text: response.data.todo,
-        completed: response.data.completed,
-      };
-    } catch (error: any) {
-      return rejectWithValue(error.response?.data?.message || 'Failed to add todo');
-    }
-  },
-);
-
-export const toggleTodoAsync = createAsyncThunk(
-  'todos/toggleTodoAsync',
-  async (
-    { id, completed, text }: { id: string; completed: boolean; text: string },
-    { rejectWithValue },
-  ) => {
-    try {
-      const response = await axios.put(`/todos/${id}`, {
-        completed: !completed,
-        todo: text,
-        id,
-      });
-
-      console.log('Toggle API Response:', response.data);
-      return {
-        id,
-        completed: !completed,
-      };
-    } catch (error: any) {
-      return rejectWithValue(error.response?.data?.message || 'Failed to toggle todo');
-    }
-  },
-);
-
-export const deleteTodoAsync = createAsyncThunk(
-  'todos/deleteTodoAsync',
-  async (id: string, { rejectWithValue }) => {
-    try {
-      const response = await axios.delete(`/todos/${id}`);
-      console.log('Delete API Response:', response.data);
-      return id;
-    } catch (error: any) {
-      return rejectWithValue(error.response?.data?.message || 'Failed to delete todo');
-    }
-  },
-);
-
-export const editTodoAsync = createAsyncThunk(
-  'todos/editTodoAsync',
-  async ({ id, text }: { id: string; text: string }, { rejectWithValue }) => {
-    try {
-      const response = await axios.put(`/todos/${id}`, {
-        todo: text,
-        id,
-      });
-
-      return {
-        id,
-        text: response.data.todo,
-      };
-    } catch (error: any) {
-      return rejectWithValue(error.response?.data?.message || 'Failed to edit todo');
-    }
-  },
-);
 
 const todoSlice = createSlice({
   name: 'todo',
@@ -128,92 +27,56 @@ const todoSlice = createSlice({
   extraReducers: (builder) => {
     builder
       // Fetch Todos
-      .addCase(fetchTodos.pending, (state) => {
-        state.status = 'loading';
-        state.loadingStates['fetch'] = true;
-        delete state.errors['fetch'];
-      })
+      .addCase(fetchTodos.pending, (state) => handlePending(state, 'fetch'))
       .addCase(fetchTodos.fulfilled, (state, action) => {
-        state.status = 'succeeded';
+        handleFulfilled(state, 'fetch');
         state.todos = action.payload;
-        state.loadingStates['fetch'] = false;
-        delete state.errors['fetch'];
       })
-      .addCase(fetchTodos.rejected, (state, action) => {
-        state.status = 'failed';
-        state.loadingStates['fetch'] = false;
-        state.errors['fetch'] = action.payload as string;
-      })
+      .addCase(fetchTodos.rejected, (state, action) =>
+        handleRejected(state, 'fetch', action.payload as string),
+      )
       // Add Todo
-      .addCase(addTodoAsync.pending, (state) => {
-        state.loadingStates['add'] = true;
-        delete state.errors['add'];
-      })
+      .addCase(addTodoAsync.pending, (state) => handlePending(state, 'add'))
       .addCase(addTodoAsync.fulfilled, (state, action) => {
+        handleFulfilled(state, 'add');
         state.todos.push(action.payload);
-        state.loadingStates['add'] = false;
-        delete state.errors['add'];
       })
-      .addCase(addTodoAsync.rejected, (state, action) => {
-        state.loadingStates['add'] = false;
-        state.errors['add'] = action.payload as string;
-      })
+      .addCase(addTodoAsync.rejected, (state, action) =>
+        handleRejected(state, 'add', action.payload as string),
+      )
       // Toggle Todo
-      .addCase(toggleTodoAsync.pending, (state, action) => {
-        const id = action.meta.arg.id;
-        state.loadingStates[id] = true;
-        delete state.errors[id];
-      })
+      .addCase(toggleTodoAsync.pending, (state, action) => handlePending(state, action.meta.arg.id))
       .addCase(toggleTodoAsync.fulfilled, (state, action) => {
+        handleFulfilled(state, action.payload.id);
         const todo = state.todos.find((todo) => todo.id === action.payload.id);
-        if (todo) {
-          todo.completed = action.payload.completed;
-        }
-        state.loadingStates[action.payload.id] = false;
-        delete state.errors[action.payload.id];
+        if (todo) todo.completed = action.payload.completed;
       })
-      .addCase(toggleTodoAsync.rejected, (state, action) => {
-        const id = action.meta.arg.id;
-        state.loadingStates[id] = false;
-        state.errors[id] = action.payload as string;
-      })
+      .addCase(toggleTodoAsync.rejected, (state, action) =>
+        handleRejected(state, action.meta.arg.id, action.payload as string),
+      )
       // Delete Todo
-      .addCase(deleteTodoAsync.pending, (state, action) => {
-        const id = action.meta.arg;
-        state.loadingStates[`delete_${id}`] = true;
-        delete state.errors[`delete_${id}`];
-      })
+      .addCase(deleteTodoAsync.pending, (state, action) =>
+        handlePending(state, `delete_${action.meta.arg}`),
+      )
       .addCase(deleteTodoAsync.fulfilled, (state, action) => {
-        const id = action.payload;
-        state.todos = state.todos.filter((todo) => todo.id !== id);
-        state.loadingStates[`delete_${id}`] = false;
-        delete state.errors[`delete_${id}`];
+        handleFulfilled(state, `delete_${action.payload}`);
+        state.todos = state.todos.filter((todo) => todo.id !== action.payload);
       })
-      .addCase(deleteTodoAsync.rejected, (state, action) => {
-        console.log(state, action);
-        const id = action.meta.arg;
-        state.loadingStates[`delete_${id}`] = false;
-        state.errors[`delete_${id}`] = action.payload as string;
-      })
+      .addCase(deleteTodoAsync.rejected, (state, action) =>
+        handleRejected(state, `delete_${action.meta.arg}`, action.payload as string),
+      )
       // Edit Todo
-      .addCase(editTodoAsync.pending, (state, action) => {
-        const id = action.meta.arg.id;
-        state.loadingStates[`edit_${id}`] = true;
-        delete state.errors[`edit_${id}`];
-      })
+      .addCase(editTodoAsync.pending, (state, action) =>
+        handlePending(state, `edit_${action.meta.arg.id}`),
+      )
       .addCase(editTodoAsync.fulfilled, (state, action) => {
+        handleFulfilled(state, `edit_${action.payload.id}`);
         const todo = state.todos.find((todo) => todo.id === action.payload.id);
-        if (todo) {
-          todo.text = action.payload.text;
-        }
-        state.loadingStates[`edit_${action.payload.id}`] = false;
-        delete state.errors[`edit_${action.payload.id}`];
+        if (todo) todo.text = action.payload.text;
       })
-      .addCase(editTodoAsync.rejected, (state, action) => {
-        const id = action.meta.arg.id;
-        state.loadingStates[`edit_${id}`] = false;
-        state.errors[`edit_${id}`] = action.payload as string;
-      });
+      .addCase(editTodoAsync.rejected, (state, action) =>
+        handleRejected(state, `edit_${action.meta.arg.id}`, action.payload as string),
+      );
   },
 });
